# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A starting point for building RESTful APIs in Go using the Gin framework, sqlx for PostgreSQL, and Redis for caching. Implementation follows Clean Architecture principles as described by Uncle Bob.",
        "title": "Go REST Boilerplate API",
        "termsOfService": "https://github.com/tenSunFree/luma-lang-go",
        "contact": {
            "name": "Najib Fikri",
            "url": "mailto:najibfikri26@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Returns an access token (short TTL) and a refresh token (longer TTL). Wrong password and unknown email take the same wall time to prevent user enumeration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authenticate and issue a token pair",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_datatransfers_requests.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tokens issued",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_datatransfers_responses.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Malformed JSON body",
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid email or password",
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Account not yet activated",
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Deletes the refresh-token jti from Redis so /auth/refresh rejects it. Access tokens remain valid until their natural expiry — clients should discard them on logout.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Revoke the refresh token",
                "parameters": [
                    {
                        "description": "Refresh token to revoke",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_datatransfers_requests.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logged out",
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Refresh token invalid",
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/password/forgot": {
            "post": {
                "description": "Emails a six-digit reset code to the address. Always returns 200 even if the email isn't registered, to defeat user enumeration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Issue a password-reset code",
                "parameters": [
                    {
                        "description": "Email address",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_datatransfers_requests.ForgotPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reset code queued (or email not registered — same response either way)",
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Malformed JSON body",
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/refresh": {
            "post": {
                "description": "Verifies the supplied refresh token, mints a new access+refresh pair, and revokes the old jti in Redis. Replaying an already-rotated token returns 401.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Rotate the refresh token, return a new pair",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_datatransfers_requests.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New token pair",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_datatransfers_responses.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Refresh token invalid, expired, or already revoked",
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Account no longer active",
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_tenSunFree_luma-lang-go_internal_http_datatransfers_requests.ForgotPasswordRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
...
        "github_com_tenSunFree_luma-lang-go_internal_http_datatransfers_requests.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 50
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 1
                }
            }
        },
        "github_com_tenSunFree_luma-lang-go_internal_http_datatransfers_requests.RefreshRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "github_com_tenSunFree_luma-lang-go_internal_http_datatransfers_responses.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_tenSunFree_luma-lang-go_internal_http_handlers_v1.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer access token issued by /auth/login or /auth/refresh.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}